+++ mpglib/common.c
@@ -65,7 +65,7 @@ const int tabsel_123 [2] [3] [16] = {
      {0,8,16,24,32,40,48,56,64,80,96,112,128,144,160,} }
 };
 
-const long freqs[9] = { 44100, 48000, 32000,
+const int freqs[9] = { 44100, 48000, 32000,
                         22050, 24000, 16000,
                         11025, 12000,  8000 };
 
@@ -231,7 +231,7 @@ decode_header(PMPSTR mp, struct frame *fr, unsigned long newhead)
 
     switch (fr->lay) {
     case 1:
-        fr->framesize = (long) tabsel_123[fr->lsf][0][fr->bitrate_index] * 12000;
+        fr->framesize = (int)((long) tabsel_123[fr->lsf][0][fr->bitrate_index] * 12000);
         fr->framesize /= freqs[fr->sampling_frequency];
         fr->framesize = ((fr->framesize + fr->padding) << 2) - 4;
         fr->down_sample = 0;
@@ -239,7 +239,7 @@ decode_header(PMPSTR mp, struct frame *fr, unsigned long newhead)
         break;
 
     case 2:
-        fr->framesize = (long) tabsel_123[fr->lsf][1][fr->bitrate_index] * 144000;
+        fr->framesize = (int)((long) tabsel_123[fr->lsf][1][fr->bitrate_index] * 144000);
         fr->framesize /= freqs[fr->sampling_frequency];
         fr->framesize += fr->padding - 4;
         fr->down_sample = 0;
@@ -269,7 +269,7 @@ decode_header(PMPSTR mp, struct frame *fr, unsigned long newhead)
         if (fr->bitrate_index == 0)
             fr->framesize = 0;
         else {
-            fr->framesize = (long) tabsel_123[fr->lsf][2][fr->bitrate_index] * 144000;
+            fr->framesize = (int)((long) tabsel_123[fr->lsf][2][fr->bitrate_index] * 144000);
             fr->framesize /= freqs[fr->sampling_frequency] << (fr->lsf);
             fr->framesize = fr->framesize + fr->padding - 4;
         }
@@ -308,7 +308,7 @@ getbits(PMPSTR mp, int number_of_bits)
         mp->wordpointer += (mp->bitindex >> 3);
         mp->bitindex &= 7;
     }
-    return rval;
+    return (unsigned int)rval;
 }
 
 unsigned int
@@ -329,7 +329,7 @@ getbits_fast(PMPSTR mp, int number_of_bits)
         mp->wordpointer += (mp->bitindex >> 3);
         mp->bitindex &= 7;
     }
-    return rval;
+    return (unsigned int)rval;
 }
 
 unsigned char
